+++ eulfedora/syncutil.py
@@ -37,7 +37,7 @@ logger = logging.getLogger(__name__)
 
 def sync_object(src_obj, dest_repo, export_context='migrate',
                 overwrite=False, show_progress=False,
-                requires_auth=False):
+                requires_auth=False, omit_checksums=False):
     '''Copy an object from one repository to another using the Fedora
     export functionality.
 
@@ -58,6 +58,9 @@ def sync_object(src_obj, dest_repo, export_context='migrate',
     :param requires_auth: content datastreams require authentication,
         and should have credentials patched in (currently only supported
         in archive-xml export mode)  (default: False)
+    :param omit_checksums: scrubs contentDigest -- aka checksums -- from datastreams;
+        helpful for datastreams with Redirect (R) or External (E) contexts
+        (default: False)
     :returns: result of Fedora ingest on the destination repository on
         success
     '''
@@ -101,10 +104,14 @@ def sync_object(src_obj, dest_repo, export_context='migrate',
             progress_bar=pbar, requires_auth=requires_auth,
             xml_only=(export_context == 'archive-xml'))
         # NOTE: should be possible to pass BytesIO to be read, but that is failing
-        export_data = export.object_data().getvalue()
+        export_data = export.object_data().getvalue()       
 
     else:
         raise Exception('Unsupported export context %s', export_context)
+
+    # wipe checksums from FOXML if flagged in options
+    if omit_checksums:
+        export_data = re.sub(r'<foxml:contentDigest.+?/>', '', export_data)
 
     dest_obj = dest_repo.get_object(src_obj.pid)
     if dest_obj.exists:
@@ -154,7 +161,7 @@ class ArchiveExport(object):
 
     #: regular expression used to identify datastream version information
     #: that is needed for processing datastream content in an archival export
-    dsinfo_regex = re.compile(r'ID="(?P<id>[^"]+)".*.*CREATED="(?P<created>[^"]+)".*MIMETYPE="(?P<mimetype>[^"]+)".*SIZE="(?P<size>\d+)".* TYPE="(?P<type>[^"]+)".*DIGEST="(?P<digest>[0-9a-f]*)"',
+    dsinfo_regex = re.compile(r'ID="(?P<id>[^"]+)".*CREATED="(?P<created>[^"]+)".*MIMETYPE="(?P<mimetype>[^"]+)".*SIZE="(?P<size>\d+)".*TYPE="(?P<type>[^"]+)".*DIGEST="(?P<digest>[0-9a-f]*)"',
         flags=re.MULTILINE|re.DOTALL)
     # NOTE: regex allows for digest to be empty
 
@@ -203,7 +210,7 @@ class ArchiveExport(object):
             self._iter_content = self.get_export().iter_content(self.read_block_size)
 
         if self._current_chunk is not None:
-            self.end_of_last_chunk = self._current_chunk[-200:]
+            self.end_of_last_chunk = self._current_chunk[-400:]
 
         self._current_chunk = self._chunk_leftover + six.next(self._iter_content)
 
@@ -249,13 +256,28 @@ class ArchiveExport(object):
             or None if no match is found
         '''
         # we only need to look at the end of this section of content
-        dsinfo = dsinfo[-250:]
+        dsinfo = dsinfo[-400:]
         # if not enough content is present, include the end of
         # the last read chunk, if available
-        if len(dsinfo) < 250 and self.end_of_last_chunk is not None:
+        if len(dsinfo) < 400 and self.end_of_last_chunk is not None:
             dsinfo = self.end_of_last_chunk + dsinfo
 
-        infomatch = self.dsinfo_regex.search(force_text(dsinfo))
+        # force text needed for python 3 compatibility (in python 3
+        # dsinfo is bytes instead of a string)
+        try:
+            text = force_text(dsinfo)
+        except UnicodeDecodeError as err:
+            # it's possible to see a unicode character split across
+            # read blocks; if we get an "invalid start byte" unicode
+            # decode error, try converting the text without the first
+            # character; if that's the problem, it's not needed
+            # for datastream context
+            if 'invalid start byte' in force_text(err):
+                text = force_text(dsinfo[1:])
+            else:
+                raise err
+
+        infomatch = self.dsinfo_regex.search(text)
         if infomatch:
             return infomatch.groupdict()
 
@@ -305,7 +327,8 @@ class ArchiveExport(object):
                 else:
                     # error if datastream info is not found, because either
                     # size or version date is required to handle content
-                    raise Exception('Failed to find datastream information from \n%s' % previous_section)
+                    raise Exception('Failed to find datastream information for %s from \n%s' \
+                        % (self.obj.pid, previous_section))
 
                 if self.xml_only and not dsinfo['mimetype'] == 'text/xml':  # possibly others?
                     try: